Verbs in classes names renamed to nouns
