+++ src/test/java/firstsimpletasks/MaxMinAverageTests.java
@@ -1,4 +1,4 @@
-package firstSimpleJavaTasks;
+package firstsimpletasks;
 
 import org.junit.Assert;
 import org.junit.Test;
